ref(profiling): make a reusable useTraces hook
